move redis handling in separate module
