+++ atciss/app/controllers/airspace.py
@@ -8,7 +8,7 @@ from pydantic import TypeAdapter
 
 from atciss.app.controllers.auth import get_user
 from atciss.app.models import User
-from atciss.app.utils import Redis, get_redis
+from atciss.app.utils.redis import Redis, get_redis
 from atciss.app.views.sector import Airport, Airspace, Position, SectorData
 from atciss.config import settings
