+++ atciss/app/controllers/metar.py
@@ -10,7 +10,7 @@ from metar.Metar import ParserError
 
 from atciss.app.controllers.auth import get_user
 from atciss.app.models import User
-from atciss.app.utils import Redis, get_redis
+from atciss.app.utils.redis import Redis, get_redis
 from atciss.app.views.metar import AirportIcao, MetarModel
 
 router = APIRouter()
