+++ atciss/app/controllers/notam.py
@@ -10,8 +10,8 @@ from sqlmodel import delete, select
 
 from atciss.app.controllers.auth import get_user
 from atciss.app.models import User
-from atciss.app.utils import Redis, get_redis
 from atciss.app.utils.db import get_session
+from atciss.app.utils.redis import Redis, get_redis
 from atciss.app.views.notam import NotamModel, NotamSeen
 
 router = APIRouter()
