+++ atciss/app/controllers/taf.py
@@ -7,7 +7,7 @@ from fastapi import APIRouter, Depends, HTTPException, Query
 
 from atciss.app.controllers.auth import get_user
 from atciss.app.models import User
-from atciss.app.utils import Redis, get_redis
+from atciss.app.utils.redis import Redis, get_redis
 from atciss.app.views.metar import AirportIcao
 
 router = APIRouter()
