+++ atciss/app/controllers/traffic.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, HTTPException
 
 from atciss.app.controllers.auth import get_user
 from atciss.app.models import User
-from atciss.app.utils import Redis, get_redis
+from atciss.app.utils.redis import Redis, get_redis
 from atciss.app.views.vatsim import AerodromeTraffic
 
 router = APIRouter()
