+++ atciss/app/utils/__init__.py
@@ -2,19 +2,9 @@
 
 import aiohttp
 from aiohttp import ClientSession
-from redis.asyncio import ConnectionPool, Redis
-
-from atciss.config import settings
-
-redis_pool = ConnectionPool.from_url(str(settings.REDIS_URL), decode_responses=True)
 
 
-async def get_redis():
-    async with Redis(connection_pool=redis_pool) as client:
-        yield client
-
-
-def aiohttp_client_session():
+def create_aiohttp_client_session():
     return ClientSession(
         timeout=aiohttp.ClientTimeout(total=60),
         connector=aiohttp.TCPConnector(
@@ -24,8 +14,5 @@ def aiohttp_client_session():
 
 
 async def get_aiohttp_client():
-    async with aiohttp_client_session() as client:
+    async with create_aiohttp_client_session() as client:
         yield client
-
-
-__all__ = ("Redis", "get_aiohttp_client", "get_redis")
