+++ atciss/tasks/aliases.py
@@ -4,7 +4,8 @@ from aiohttp import ClientConnectorError, ClientSession
 from fastapi import Depends
 from loguru import logger
 
-from atciss.app.utils import Redis, get_aiohttp_client, get_redis
+from atciss.app.utils import get_aiohttp_client
+from atciss.app.utils.redis import Redis, get_redis
 from atciss.config import settings
 from atciss.tkq import broker
