+++ atciss/tasks/basic_ad.py
@@ -5,7 +5,8 @@ from fastapi import Depends
 from loguru import logger
 from pydantic import RootModel, TypeAdapter
 
-from atciss.app.utils import Redis, get_aiohttp_client, get_redis
+from atciss.app.utils import get_aiohttp_client
+from atciss.app.utils.redis import Redis, get_redis
 from atciss.app.views.basic_ad import BasicAD
 from atciss.tkq import broker
