+++ atciss/tasks/taf_metar.py
@@ -6,7 +6,8 @@ from aiohttp import ClientConnectorError, ClientSession
 from fastapi import Depends
 from loguru import logger
 
-from atciss.app.utils import Redis, get_aiohttp_client, get_redis
+from atciss.app.utils import get_aiohttp_client
+from atciss.app.utils.redis import Redis, get_redis
 from atciss.tkq import broker
